cros: x86: Add 2MB RW section for SeaBIOS payload
In order to allow u-boot to chain load a SeaBIOS payload
to boot Windows, we need a section in flash to place SeaBIOS.
Due to limitations of the flash hardware write protection,
we can either have a 4MB or 2MB RO block. Hence, giving the
full 2MB to SeaBIOS.
This will have to change slightly when we implement verification
of the SeaBIOS payload.

BUG=chrome-os-partner:6108
TEST=emerge-link chromeos-u-boot chromeos-bootimage builds and
     produces a booting firmware image.
BRANCH=link
Signed-off-by: Stefan Reinauer <reinauer@chromium.org>

Change-Id: I5d296b1d76977c75aa71e0036c48f5161848a6e7
Reviewed-on: https://gerrit.chromium.org/gerrit/31249
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Commit-Ready: Stefan Reinauer <reinauer@google.com>
Tested-by: Stefan Reinauer <reinauer@google.com>