cros: Detect power button 'press' rather than 'hold'
If the power button is held on boot, we don't want to turn off
immediately. Record the previous state of vboot flags so that we can
detect a rising edge, i.e. an actual press of the power button rather
than it just being held since boot.

Note that on daisy the power button will always appear to be held on
boot due to the EC asserting XPSHOLD. We could potentially reduce the
1s delay to 150ms or so (from the current 1s).

BUG=chrome-os-partner:12748
BRANCH=snow
TEST=manual
Using snow, holding the power button on boot should not power off in
U-Boot:

dut-control cold_reset:on sleep:.1 cold_reset:off sleep:.1
    pwr_button:press sleep:<hold_time> pwr_button:release

where <hold_time> is .1, .5, 1, 2, 3 up to 15

For values of about 16 or more the EC will force power off.

Also boot into recovery mode and check that we can power off correctly
at a screen.

Change-Id: Ia52d59d5af809a30ae02b55077b3ded3867d79e2
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31318
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>