Fix Stumpy/Lumpy ToT
The 'second generation' legacy platforms (Stumpy/Lumpy) were left
behind and do not boot with ToT firmware anymore. With coreboot fix
supplying the option ROM status in place, all what's needed is adding
the ec-in-rw value to the device tree. With this change it is
hardcoded to true.

BRANCH=none
BUG=chrome-os-partner:13268
TEST=manual
  . build new firmware image for Lumpy
  . program it on a device
  . boot CromeOS in dev and normal mode
  . reboot in recovery mode and install ChromeOS

Change-Id: Ief33bd85f888bcb069c2c4973e26e0b68ad166d7
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31678
Reviewed-by: Stefan Reinauer <reinauer@google.com>