Make sure vboot_flag subsystem is initialized before use
It was noticed that in some cases vboot_flag attributes are attempted
to read before the infrastructure is initialized, in particular - in
cros/coreboot/keyboard.c:board_i8042_skip(). Not sure if there are
other instances now, but nothing prevents from adding vboot_flag
access functions before the infrastructure is ready.

With this change the initialization happens as soon as needed.

BRANCH=none
BUG=none
TEST=manual
  . build the new boot image and start it on Lumpy
  . observe that error message about developer switch info not
    available is not printed anymore

Change-Id: Ideb8de486a8fe9219586e1646d6e73b782b9c628
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31679
Reviewed-by: Stefan Reinauer <reinauer@google.com>