exynos: Increase i2c arbitration wait time
We currently wait 50ms for the EC to release the i2c bus. The longest
transaction we expect the EC to perform is 4 bytes, which should be
about 0.5ms, but if there is a bug in the EC driver the time could
wind out significantly.

Change the wait time to 500ms.

Once the EC I2C driver is reliable, we can revisit this if we are still
using I2C.

BUG=chrome-os-partner:13338
BRANCH=snow
TEST=manual

On AP:

cros_test i2c

On EC:

pmu 1000

See that all i2c transactions succeed.

Change-Id: I86005a2e2ee48ce04247fbf5a31a77779f3ff07a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31918
Reviewed-by: Doug Anderson <dianders@chromium.org>