gen: sandbox: cros: add chromeos_sandbox configuration
This change adds a new sandbox variant that will be used to implement
ChromeOS-specific aspects.  This is needed because ChromeOS has
additional features which are not present in the standard U-Boot
distribution, and we'd like the sandbox system to run with a standard
U-Boot distribution, as well as with the ChromeOS additions.

BRANCH=none
BUG=chromium-os:32603
TEST=Code is not used yet
Signed-off-by: Taylor Hutt <thutt@chromium.org>

Change-Id: I58b1f24483f4778aae788dad0b7618bcd69d9cfa
Reviewed-on: https://gerrit.chromium.org/gerrit/32411
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Taylor Hutt <thutt@chromium.org>
Tested-by: Taylor Hutt <thutt@chromium.org>