:arm :i2c Driver handles STOP event errors
Previously, the STOP bits were sent without checking if it was
successful or not.  Since the STOP bit needs to get sent even if there
was a transmission error (you don't want to leave a slave device
hanging, waiting for the bus to clear) it was overlooked.  Now if the
transmission is successful, but the STOP bit fails the driver will
return a timeout.  However, if the transmission fails, we really don't
care that the STOP bit timed out, we'd rather see how the actual
transmission died, so it will return that if there was an error prior to
trying to send the stop bit.

BUG=chrome-os-partner:13499
TEST=boot the machine, make sure everything powers on correctly.  From
the Uboot console run a bunch of i2c commands (mkbp read ro 0x00
in particular, really tests the i2c), everything should still work.
BRANCH=snow

Change-Id: Id6ed87ca423acd650ef28c292534677946f67b67
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32028
Reviewed-by: Doug Anderson <dianders@chromium.org>