exynos: clock: Fix a typo bug in exynos clock init
We intended to clear the bits of CLK_SRC_TOP2 register, instead we were
writing on the reserved bits of src_core1 register. Since the default
value of clk_src_top2 register were itself zero, this typo was not
creating any big issue. But it is better to fix this error for better
readability of the code.

BRANCH=snow
BUG=chrome-os-partner:13794
TEST=build ok. Booted till the chromeos login prompt on snow.

Change-Id: I8da954e1989d03c2cf90d9084ac22052a31fa478
Signed-off-by: Hatim Ali <hatim.rv@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/32543