cros: Refactor display backup code to display properly
The existing code looks ugly on snow since the text displays across
a line in the middle of the screen.

Update the screen to display the message on its own line. Also add
the wait screen which was missing.

BUG=chrome-os-partner:13492
BRANCH=snow
TEST=manual
Run s/w sync with missing images and see that it now displays the
fallback image properly.

Change-Id: I600b0d57c42cbd3720c7126e702bd40be6b8a9eb
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32574
Reviewed-by: Mike Truty <truty@chromium.org>