cros: Read bmpblk from GBB when the display is inited
If vboot calls VbExDisplayInit() then we can expect that it will start
looking through the bitmaps for one to display. Unfortunately on ARM we
are currently guessing what vboot might need from the GBB, since it takes
about 200ms to read the whole thing.

Normally we guess correctly, since we can use the VB_INIT_OUT_ENABLE_DISPLAY
flag to tell us that the display will be needed. This doesn't work for
software sync, however, so we need to copy with an unexpected call to
VbExDisplayInit().

This is a work-around. When we come to clean this up we should remove
CONFIG_HARDWARE_MAPPED_SPI and unify the code paths, as well as change vboot
to request data from GBB before it uses it.

BUG=chrome-os-partner:13492
BRANCH=snow
TEST=manual
Reflash snow with vboot modified to always software sync. Update U-Boot fdt
to set preamble-flags to 0. Remove the factory-friendly flags so that
developer mode is not forced (not strictly necessary).

Do a normal boot and see that the wait screen now appears on snow correctly.

Change-Id: Iae2e34c485383272d4feb1e90c22bf9a4d835ce1
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32632
Reviewed-by: Randall Spangler <rspangler@chromium.org>