Do not handle bitmaps which do not fit into the screen
The cfb console driver is trying to prevent bitmaps to spill over the
screen, but the calculations assume that at least part of the bitmap
fits into the screen area. In reality there could be bitmap elements
which are completely out of the screen area, they just need to be
discarded.

BRANCH=link
BUG=chrome-os-partner:13876
TEST=manual
  . build a link image
  . replace the bitmap with a production varian using gbb_utility:
    sudo gbb_utility -s --bmpfv=../s1280a8x5-1.0.0.bin \
       /build/link/firmware/image-link.bin --flags=0
  . load the new image on a Link and boot it in recovery mode
  . observe that the image on the screen is no longer corrupted

Change-Id: Ifc9ce531276b6d0546f48716187b38ca14451a02
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32886
Reviewed-by: Stefan Reinauer <reinauer@google.com>