gen: Add stricmp() and strnicmp()
strnicmp() is present but disabled. Make it available and define stricmp()
also. There is a only a small performance penalty to having stricmp()
call strnicmp(), so do this instead of a standalone function, to save code
space.

BUG=chrome-os-partner:14098
BRANCH=none
TEST=manual
Build U-Boot and see that it boots on snow.
Build for link.

Change-Id: I38a0cfaa1d420dc3542f1137ef07882f7e6bb34b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33330