exynos: Add use_bpll option to memory parameters
Some memory speeds want to use the BPLL and some don't. Add a parameter to
take care of this, to avoid special cases in the code for each memory
speed.

BUG=chrome-os-partner:14151
BRANCH=none
TEST=manual
build and boot on snow, see that U-Boot still comes up
Boot into Linux and login as guest. Browse to a few web sites.

Change-Id: I74e04a0ba7761992aecf890db7dac8c0fa4eed03
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33918
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-by: Hatim Ali R.V <hatim.rv@samsung.com>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>