Revert "input: Add ANSI 3.64 escape sequence generation."
I have taken the rather unusual step of reverting the old commit
and creating a new one. I am sorry that this will make it a bit harder
to review, but I would like to send the new commit upstream, and this
will make it easier for a rebase later.

Sorry for the trouble.

This reverts commit 8462d52d1da5f868a408f80d609faf9587873bf3.

BUG=chrome-os-partner:13864
BRANCH=none
TEST=none
(this actually breaks arrow keys)

Change-Id: I4f60a2aac87a0ae9fa7c683bed5f5d7dc4a20288
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33961