gen: mkbp: Read keyscans until we get a key or there are no more
We need tstc() to scan all available key scans looking for a keypress.
Once it finds one, it is ok to exit, since we know that another call
to tstc() will do the same.

tstc() relies on mkbp_kbc_check() to check for key scans, so add a
loop there to cope with this case.

BUG=chrome-os-partner:13864
BRANCH=none
TEST=manual
Instrument do_vboot_twostop() to detect a tstc() that does the wrong thing.

	/*
	 * Empty keyboard buffer before boot.  In case EC did not clear its
	 * buffer between power cycles, this prevents vboot of current power
	 * cycle being affected by keystrokes of previous power cycle.
	 */
	while (tstc())
		getc();

	/* Test that there really are no pending characters! */
	{
		int i;

		for (i = 0; i < 20; i++) {
			if (tstc()) {
				printf("oh dear\n");
				while (1)
					;
			}
		}
	}

Run the code without this CL and see that we get a failure. With this change
we don't.

Tested on snow and link.

Change-Id: Idbbd4b004b3aa86d91f05822699a22cc14ffca06
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33963