Make the FMAP available to VbExProtectFlash()
In order to temporarily write-protect the RW BIOS that we're about to
execute, this happens:

twostop calls VbSelectFirmware()
  VbSelectFirmware() calls VbExProtectFlash()
    VbExProtectFlash() needs to find the range of these FMAP sections:
      RW_SECTION_A
      RW_SECTION_B

While the first two regions are available in the struct twostop_fmap
fmap, that is a local variable inside the twostop functions, and is not
 passed or available to VbExProtectFlash().

This CL make it available internal to the file.

BUG=chrome-os-partner:13342
BRANCH=Link
TEST=none

Refactoring only, no visible change, nothing to test.

Change-Id: I7f4fb34cde0ed98789f77f7575620943f6eba8b6
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34007
Reviewed-by: Randall Spangler <rspangler@chromium.org>