cros: Add spi_write_protect_region() to write-protect flash regions.
This adds a function that uses the Protected Range Registers to make regions
of the SPI flash read-only until the next boot. We'll use this to prevent
changes to the running RW BIOS copy.

This change just adds the function. A future change will make use of it.

BUG=chrome-os-partner:11215
BRANCH=Link
TEST=none

This just adds an internal function. Nothing calls it yet, so nothing to test.

Change-Id: Ib547139be6f0437f4e0edbdb5242d6d7e3fb4211
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34057
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>