Pass VB_INIT_FLAG_SW_WP_ENABLED value to VbInit()
This adds a function to the SPI flash interface so that flash components
that support it can determine whether or not parts of the flash are
protected via software settings in the part itself.

If implemented and true, U-Boot will pass this information to VbInit().

This CL only implements the driver for the Winbond SPI flash.

BUG=chrome-os-partner:13266
BRANCH=none (this needs thinking about)
TEST=manual

To test on Link, be sure to use the real Winbond BIOS flash, not an EM100
pretending to be one.

1. Take the D-panel off and remove the HW WP screw (the 1/4" copper washer
   on the left-hand edge of the motherboard, between the USB jack and the
   speaker).

   Boot and run

     crossystem | grep wpsw_boot

   You should see this:

     sw_wpsw_boot       = 0
     wpsw_boot          = 0

2. Replace the HW WP screw, reboot, run the command again. Now you should
   see:

     sw_wpsw_boot       = 0
     wpsw_boot          = 1

3. Run this:

     flashrom --wp-range 0x600000 0x200000
     flashrom --wp-enable

   Reboot, and check again. You should see this:

     sw_wpsw_boot       = 1
     wpsw_boot          = 1

4. Remove the HW WP screw and reboot once more. Now it should be back to the
   values in step 1:

     sw_wpsw_boot       = 0
     wpsw_boot          = 0

Change-Id: Ie87b5b76c74c0b81c0d3ed3a73c60d11d14a63df
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34895
Reviewed-by: Randall Spangler <rspangler@chromium.org>