gen: video: Check for valid FB pointer before clearing
This command will start erasing at memory address zero
if there is not a valid framebuffer address that was found
during video_init().

This is a common case with Chrome OS devices in normal mode
when we do not execute the video option rom in coreboot.

BUG=chrome-os-partner:15024
BRANCH=link,stout,butterfly
TEST=manual: complete multiple successful boots in normal mode

Change-Id: I32c5c984ce2e45142f9452d14492ed94fa398e82
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34931
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>