x86: config: Define CONFIG_OF_CBFS and CONFIG_OF_SEPARATE in coreboot
We currently use CBFS to obtain the fdt. We are adding option for this,
so define it to ensure continued operation. We will move away from this
once crosbug.com/p/15502 is done.

For RW firmware we want to use CONFIG_OF_SEPARATE as on ARM, so that it
can have its own fdt.

BUG=chrome-os-partner:14545
BRANCH=link
TEST=build and boot on link, type 'vboot_twostop' see that it boots
correctly.

Change-Id: I9b6da952734a3689a1e91ca4e7c54e05c6905f48
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/36204
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>