x86: coreboot: Make CONFIG_OF_CBFS an option
We want to stop using this feature on coreboot. In preparation for this,
make it optional, and even if CONFIG_OF_CBFS is defined, we only do it
on the cold boot.

Also move the serial setup into a separate function, board_init_f(), so that
we can do this even if the fdt comes from CONFIG_OF_SEPARATE.

BUG=chrome-os-partner:14545
BRANCH=link
TEST=build and boot on link, type 'vboot_twostop' see that it boots
correctly.

build with and without CONFIG_OF_CBFS.

Change-Id: Id4ec8991c736c971a043589b1a4d789bb76566e6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/36206
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>