cros: fdt: Use separate RW-A/B fdt for coreboot boards
We can now select a different fdt for RW A/B firmware. Use this for both
EC and non-EC builds.

BUG=chrome-os-partner:14545
BRANCH=link
BUG=chrome-os-partner:14545
BRANCH=link
TEST=boot into U-Boot with pre-emable flags set to 0.

See that it indicates RO firmware:

> vboot_test fmap
...
Firmware type: 0

Then change to RW-A firmware:

> vboot_twostop
...

Once in RW-A firmware:

> vboot_test fmap
...
Firmware type: 1

Note: RW-B has not been tested. This can be done by modifying vboot to choose
B instead of A.

Change-Id: Id25a6ce871c05c8fb4410e9f5abc005698c202ab
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/36208
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>