split: cros: exynos: x86: Add mkbp error handling
When the MKBP device does not init properly, it should return an error.
This is distinct from the case where we don't have an MKBP fdt node and
thus don't expect the device to be there.

When MKBP init fails, we need to stop. On snow, print a message and halt,
for now. On link, panic(), which will cause a reset.

BUG=chrome-os-partner:13940
BRANCH=none
TEST=manual
Build on snow with mkbp_init() changed to fail. See that we get a message
on the display and the machine hangs.

Build on link with mkbp_init() changed to fail. See that we get a message
on the uart console. However the reset does not appear to work: __reset_cpu()
just hangs.

Change-Id: Ie1e200e6206ee1205cd184122c03b7dcb04f3531
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/36350
Reviewed-by: Randall Spangler <rspangler@chromium.org>