fdt: add Analogix bridge on Spring
The Analogix ANX7805 is either a myDP compliant external output or a USB
pass-through. At startup, we want to use the port as USB to be able to
boot from it. So put the ANX7805 is pass-through mode (its default
operating mode) by getting it out of power-down and reset.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=chrome-os-partner:14390
TEST=on Spring, boot from a USB stick on USB port 0.

Change-Id: If021515d543fec788515336195bf0a4023337468
Reviewed-on: https://gerrit.chromium.org/gerrit/36924
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>