cros: Add a FAFT servo keyboard remapping via fdt config
This change supports two arrays of key pairs, each able remap three keys.
The first is for normal ascii keys, which are a one to one mapping.
The second is for special keys, which have the ESC and '[' character
already stripped off.

BUG=chrome-os-partner:14056
TEST=none
BRANCH=stout

Change-Id: Iab305a4535a649f74d7df29f98c51b0219df5da7
Signed-off-by: Marc Jones <marc.jones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/37078
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>