crossystem: Tidy up the fdt code to report errors properly
We need to fail if anything goes wrong with the fdt code here.
Report the error on the line where it happened. This adds to code
size but makes it easier to debug.

BUG=chromium-os:35959
BRANCH=none
TEST=build for snow
boot into kernel, look at /proc/device-tree/firmware/chromeos and
see that things look normal.
build for link

Change-Id: Ia6cafb26de782d0dd6a4a33647538ad9103e6806
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/37172
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>