upstream: patman: Add support for settings in .patman
This patch adds support for a [settings] section in the .patman file.
In this section you can add settings that will affect the default
values for command-line options.

Support is added in a generic way such that any setting can be updated
by just referring to the "dest" of the option that is passed to the
option parser.  At the moment options that would make sense to put in
settings are "ignore_errors", "process_tags", and "verbose".  You
could override them like:

 [settings]
 ignore_errors: True
 process_tags: False
 verbose: True

The settings functionality is also used in a future change which adds
support for per-project settings.

BUG=None
TEST=Use patman for upstreaming
BRANCH=none

NOTE: not actually upstream, but posted and acked.
http://patchwork.ozlabs.org/patch/203519/

Change-Id: Ie4e165b974431e970cb637df8dfc1cb399224ecf
Signed-off-by: Doug Anderson <dianders@chromium.org>
Acked-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38579
Reviewed-by: Simon Glass <sjg@chromium.org>