upstream: patman: Add a call to get_maintainer.pl if it exists
For Linux the best way to figure out where to send a patch is with the
"get_maintainer.pl" script.  Add support for calling it from patman.
Support is added unconditionally for "scripts/get_maintainer.pl" in
case it is helpful for any other projects.

BUG=None
TEST=Use patman for Linux work; see that get_maintainer.pl is called.
BRANCH=none

NOTE: not actually upstream, but posted and acked.
http://patchwork.ozlabs.org/patch/203520/

Change-Id: Ic0ddb79c96928cbcb7d710537dfc499ba26c6a5a
Signed-off-by: Doug Anderson <dianders@chromium.org>
Acked-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38689
Reviewed-by: Simon Glass <sjg@chromium.org>