drop: i8042: Fix Caps Lock behaviour
The i8042 code seems to deal with Caps Lock incorrectly, in that it
processes it as a numeric keypad key. Since it is not in range this
can cause an out-of-bound array access to us_kbd_numkey*[].

BUG=chrome-os-partner:16347
BRANCH=butterfly,stout,parrot
TEST=none yet
Need to test on a machine to make sure this fixes the bug.

Change-Id: I97a346098837b8fd992878100b84e3ceec523d0b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38672
Reviewed-by: Shawn Nematbakhsh <shawnn@google.com>
Tested-by: Shawn Nematbakhsh <shawnn@google.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>