upstream: patman: Cache the CC list from MakeCcFile() for use in ShowActions()
Currently we go through and generate the CC list for patches twice.
This gets slow when (in a future CL) we add a call to
get_maintainer.pl on Linux.  Instead of doing things twice, just cache
the CC list when it is first generated.

BUG=None
TEST=Use patman for upstreaming
BRANCH=none

NOTE: not actually upstream, but posted and acked.
http://patchwork.ozlabs.org/patch/203521/

Change-Id: I5bb3a86cdd782d65eb3cf979685cc953312edcde
Signed-off-by: Doug Anderson <dianders@chromium.org>
Acked-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38862
Reviewed-by: Simon Glass <sjg@chromium.org>