cros: Take the offset of the "boot" subsection into account.
The RW firmware on Link is actually composed of several pieces, the system or
"boot" firmware, the EC RW, and the RW device tree. This change leaves in
place the assumption that the boot part is first but starts looking at the
contents of the RW portion where boot actually starts, rather than just the
beginning. This allows having something else at the start that U-Boot can
ignore but would describe the layout to a firmware not using the device tree.

BUG=chrome-os-partner:16412
TEST=Built and booted on Link with and without something preceeding the system
firmware.
BRANCH=None

Change-Id: Ia85fad61b6ceb92dc097986c2d0a8b33236af107
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/39308
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>