vboot: Add some protection against invalid hashes
VbExEcHashRW() now waits for the EC to finish if it's busy calculating a
hash, and if the hash it calculates is wrong in some way, it kicks off
another attempt to try to get it right.

VbExEcUpdateRW() will tell the EC to invalidate the current hash before it
begins replacing the RW firmware, so that if the AP is interrupted the next
time around the hash will be invalid instead of just matching what used to
be there.

BUG=chrome-os-partner:16668
BRANCH=all
TEST=manual

Testing is a long involved process when done manually. The quickest way is
to run the firmware_SoftwareSync FAFT test that uncovered the problem in the
first place.

Change-Id: I98aa8de8a8fee94534823d3bc84b11fd9d488cf4
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39526
Reviewed-by: Randall Spangler <rspangler@chromium.org>