exynos: config: Place environment at the end of SPI flash
The old location for the environment used a bunch of MMC-related
defines to figure out where to place the environment in SPI flash.
This didn't make a lot of sense, though it ended up placing the
environment in a harmless location (see http://crosbug.com/37144 for
details).  Place the environment in the correct location.  This has
the side effect of allowing you to save/restore the environment even
if the beginning of SPI flash is write protected.

BUG=chromium-os:37144
BRANCH=none
TEST=Add printout to env_sf.c and see that it now prints out that the
environment is now saved to 0x003fc000
TEST=Run this U-Boot on a device with write-protect still enabled and
see that environment can be updated.

Change-Id: I200fa76cc520e19c914b8560a8c4afd38e08344c
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39550
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>