arm: Call bootstage_relocate() after malloc is initted
In a previous CL we added the bootstage_relocate(), which should be
called after malloc is initted.  Now we call it on the ARM board.
Once the generic board uses bootstage we should put it there as well.

BUG=chromium-os:37195
TEST=Run vboot and don't see blank string for the stage after
spl_start in the "Timer summary in microseconds:".  Instead see
"board_init_f"
BRANCH=snow

Change-Id: I2064990dbb4cf19558693288e2df12984d1e5f7c
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39684
Reviewed-by: Simon Glass <sjg@chromium.org>