Revert "U-boot bss needs to be initialized"
The BSS should not be used prior to relocation, since it is not set up and
does not really exit. The address we have for it is within the linked image,
but typically no space is allocated for it, and thus we may be trampling
over other contents in the image. Typically this is the relocation tables.

We certainly should not zero it, since this wipes our relocation tables.

BUG=chrome-os-partner:13961
BRANCH=link
TEST=manual
build and boot to kernel on link

This reverts commit e350fa9df69e541229709e89003c8ecd0a4db8fc.

Change-Id: Ic542a4e98aa0c5160bba8e984c694684efd05bd8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39945
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>