x86: spi: Support SPI speed in ICH9 driver
This peripheral allows some control over SPI speed (selecting 20MHz or
33MHz). When the speed requested is above 33MHz, select 33MHz.

The control masking used in this driver is a bit wrong, since it is using
a 16-bit access but worrying about buts 23:16. Also it should probably
read the old control register value so as not to affect bits that need
to stay the same.

The only way to write to bits 23:16 of the control register seems to be
with a writeb(), since a writel() to an unaligned address doesn't work
(the control register is at an unaligned offset of 0x91 for some reason).

BUG=chrome-os-partner:13961
BRANCH=link
TEST=manual
Try at 20MHz and 33MHz and see if we can see a time difference

sf probe 0 20000000
time sf read 100000 0 800000

sf probe 0 33000000
time sf read 100000 0 800000

Some of the time the faster speed completes in about 4 seconds and the
slower speed in about 5 seconds. That's not really a conclusive test but
from what I understand there are other uses of the SPI bus that we cannot
control.

Change-Id: I85ec8c2f70fb697b46cfe8e26a1ce3744f655610
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39947
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>